SN2005ks's claimedtype from 2014arXiv1401.3317S,2012A&A...538A.120L,2008yCat....1.2024B,2008AJ....135..348S,2000A&AS..143....9W,1989A&AS...81..421B marked as being erroneous.
this is  classified as a Ic hypernova, i.e. a Ic-BL based on the similarity to 02ap (Table 3 Zeng+08)  http://iopscience.iop.org/article/10.1088/0004-6256/135/5/1766/meta
Note that Zeng+08 uses "Ic" for bothregularIc and Ic-BL. The latter however shows the template 02ap and appears as Ic(02ap)/hyper in Table 4. There may be others...