SN2005by's claimedtype from 2008yCat....1.2024B,2005IAUC.8523....4R,Latest Supernovae,2000A&AS..143....9W,1989A&AS...81..421B marked as being erroneous.
missing the original reference for the IIb typing which is http://adsabs.harvard.edu/abs/2005ATel..509....1G. I set the alias to 5 as it should be chronologically. I dont know if I should report the references with ;arger aliases as "erroneous" or if I should not try to set the alias.